TSK-3620, TSK-3621, TSK-3622, TSK-3623, TSK-3625, TSK-3626.

+++ database/migrations/2024_06_14_134947_create_car_rental_has_cars_table.php
@@ -31,7 +31,8 @@ return new class extends Migration
             $table->integer("number_seats");
             $table->string('category_rent');
             $table->enum("category", ["manual", "automatic"]);
-            $table->float("rental_price_per_day");
+            // $table->float("rental_price_per_day");
+            $table->decimal('rental_price_per_day', 15, 2);
             $table->string('duration');
             $table->string("status");
             $table->text("description");
